+++ OUDS/Core/Themes/Wireframe/Tests/WireframeThemeTests.swift
@@ -110,6 +110,10 @@ struct WireframeThemeTests {
         #expect(theme.controlItem is WireframeThemeControlItemComponentTokensProvider)
     }
 
+    @Test func defaultIconComponentTokensProvider() throws {
+        #expect(theme.icon is WireframeThemeIconComponentTokensProvider)
+    }
+
     @Test func defaultLinkComponentTokensProvider() throws {
         #expect(theme.link is WireframeThemeLinkComponentTokensProvider)
     }